Fixed uninitialized variable comparison on nonexistent ABS packages

+++ PlugApps/Build/Database.pm
@@ -426,11 +426,12 @@ sub update {
 				}
 				next;
 			}
-			# new package, different version, update, done = 0
-			next unless (! defined $db_pkgver);
+			# skip a bad source
+			next if (! defined $pkgver);
 			# create work unit here, to repackage abs changes without ver-rel bump
 			`tar -zcf "$workroot/$repo-$pkg.tgz" -C "$absroot/$repo" "$pkg" > /dev/null`;
-			next unless ("$pkgver-$pkgrel" ne "$db_pkgver-$db_pkgrel");
+			# new package, different version, update, done = 0
+			next unless (! defined $db_pkgver || "$pkgver-$pkgrel" ne "$db_pkgver-$db_pkgrel");
 			print "$repo/$pkg to $pkgver-$pkgrel\n";
 			$self->{dbh}->do("delete from package where package = '$pkg'");
 			$self->{dbh}->do("insert into package (package, repo, pkgname, provides, pkgver, pkgrel, depends, makedepends, git, abs) values (?, ?, ?, ?, ?, ?, ?, ?, 0, 1)",
